Remove old operator function sand operate functions and add new functions.
Add new calculate, addDigit, addOperator, decimalPoint, getResult functions,
and clearInputScreen functions.

Add global firstNum, secondNum, and round variables
The round variable keeps says wether the current calculation was the
first or not.

This changed method works so that the first lot of numbers inputted into
lowerInput is == firstNum and after the operator is added the lowerInput
gets pushed up to the upperInput along with the operator and what gets
put into lowerInput next is == secondNum. When equals is clicked the lowerInput
is added to the upperInput with the previous text and the result of the
calculate function is put into the lowerInput, then round is incremented.
else if round !== 1 but operator == true
The lowerInput which would be the result of the last calculation is
pushed up to upperInput with the chosen operator and secondNum. Then the
calculate will run.